requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1
qdrant-client==1.7.0
pinecone-client==2.2.4

# ML & AI
transformers==4.36.0
onnxruntime==1.16.3
optimum==1.16.0
sentence-transformers==2.2.2
rank-bm25==0.2.2
openai==1.6.1

# NLP
langchain==0.1.0
langchain-community==0.0.10

# Evaluation
ragas==0.1.0
rouge-score==0.1.2
datasets==2.16.0

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6
bcrypt==4.1.2
PyYAML==6.0.1
numpy==1.24.3